+++ bootpay_backend/commerce/modules/__init__.py
@@ -16,6 +16,13 @@ from .cart import CartModule
 from .order_subscription_request import OrderSubscriptionRequestModule
 from .mall_setting import MallSettingModule
 from .webhook import WebhookModule
+from .alimtalk_message import AlimtalkMessageModule
+from .alimtalk_official import AlimtalkOfficialModule
+from .alimtalk_optout import AlimtalkOptoutModule
+from .alimtalk_send import AlimtalkSendModule
+from .alimtalk_sender import AlimtalkSenderModule
+from .alimtalk_template import AlimtalkTemplateModule
+from .alimtalk_webhook import AlimtalkWebhookModule
 
 __all__ = [
     'UserModule',
@@ -34,5 +41,12 @@ __all__ = [
     'CartModule',
     'OrderSubscriptionRequestModule',
     'MallSettingModule',
-    'WebhookModule'
+    'WebhookModule',
+    'AlimtalkMessageModule',
+    'AlimtalkOfficialModule',
+    'AlimtalkOptoutModule',
+    'AlimtalkSendModule',
+    'AlimtalkSenderModule',
+    'AlimtalkTemplateModule',
+    'AlimtalkWebhookModule'
 ]